support frontend only mode
