+++ source/common/types/cardano-node.types.js
@@ -1,5 +1,6 @@
 // @flow
 export type TlsConfig = {
+  hostname: string,
   port: number,
   ca: Uint8Array,
   cert: Uint8Array,
