+++ source/renderer/app/api/common/types.js
@@ -1,5 +1,6 @@
 // @flow
 export type RequestConfig = {
+  hostname: string,
   port: number,
   ca: Uint8Array,
   cert: Uint8Array,
