+++ source/renderer/app/api/nodes/requests/applyNodeUpdate.js
@@ -6,7 +6,6 @@ export const applyNodeUpdate = (
   config: RequestConfig,
 ): Promise<any> => (
   request({
-    hostname: 'localhost',
     method: 'POST',
     path: '/api/internal/apply-update',
     ...config,
