+++ source/renderer/app/api/nodes/requests/getNextNodeUpdate.js
@@ -6,7 +6,6 @@ export const getNextNodeUpdate = (
   config: RequestConfig
 ): Promise<any> => (
   request({
-    hostname: 'localhost',
     method: 'GET',
     path: '/api/internal/next-update',
     ...config,
