+++ source/renderer/app/api/nodes/requests/postponeNodeUpdate.js
@@ -6,7 +6,6 @@ export const postponeNodeUpdate = (
   config: RequestConfig
 ): Promise<any> => (
   request({
-    hostname: 'localhost',
     method: 'POST',
     path: '/api/internal/postpone-update',
     ...config,
