+++ source/renderer/app/api/wallets/requests/importWalletAsJSON.js
@@ -8,7 +8,6 @@ export const importWalletAsJSON = (
   filePath: string,
 ): Promise<AdaWallet> => (
   request({
-    hostname: 'localhost',
     method: 'POST',
     path: '/api/backup/import',
     ...config,
