+++ source/renderer/app/api/wallets/requests/resetWalletState.js
@@ -6,7 +6,6 @@ export const resetWalletState = (
   config: RequestConfig
 ): Promise<void> => (
   request({
-    hostname: 'localhost',
     method: 'DELETE',
     path: '/api/internal/reset-wallet-state',
     ...config,
