+++ source/renderer/app/api/wallets/requests/restoreWallet.js
@@ -9,7 +9,6 @@ export const restoreWallet = (
   { walletInitData }: { walletInitData: WalletInitData }
 ): Promise<AdaWallet> => (
   request({
-    hostname: 'localhost',
     method: 'POST',
     path: '/api/v1/wallets',
     ...config,
